+++ serveur/utilisateur.h
@@ -13,9 +13,9 @@ public:
     utilisateur(QTcpSocket* socket);
     void editpseudo(QString newpseudo);
     void editversion(QString newvertion);
-    void setmessageSize(quint16 size);
+    void setmessageSize(int size);
     QTcpSocket* getSocket();
-    quint16 getmessageSize();
+    int getmessageSize();
     QString getpseudo();
     QString getversion();
     bool isconnecteed();
@@ -28,6 +28,6 @@ private:
     QString vertion;
     QString pseudo;
     QTcpSocket* socket;
-    quint16 messageSize;
+    int messageSize;
 };
 #endif // UTILISATEUR_H
